tools/requirements.txt
claptcha==0.3.3
Pillow==8.1.0
smtp==0.1.0
geoip2==4.1.0
